+++ FE/src/app/layout.tsx
@@ -1,10 +1,7 @@
 "use client";
 
-import { Providers } from "@/store/provider";
+import { Providers } from "@/store/providers";
 import "./globals.css";
-import { QueryClient, QueryClientProvider } from "react-query";
-
-const queryClient = new QueryClient();
 
 export default function RootLayout({
   children,
@@ -12,12 +9,10 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <QueryClientProvider client={queryClient}>
-      <Providers>
-        <html lang="en">
-          <body className={"bg-black"}>{children}</body>
-        </html>
-      </Providers>
-    </QueryClientProvider>
+    <Providers>
+      <html lang="en">
+        <body className={"bg-black"}>{children}</body>
+      </html>
+    </Providers>
   );
 }
