--- FE/src/store/provider.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { store } from "./store";
-import { Provider } from "react-redux";
-
-export function Providers({ children }: { children: React.ReactNode }) {
-  return <Provider store={store}>{children}</Provider>;
-}
